requirements.txt
azure-ai-formrecognizer>=3.2.0
azure-ai-vision-imageanalysis>=1.0.0
azure-ai-textanalytics>=5.2.0
azure-storage-blob>=12.16.0
azure-identity>=1.14.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pytest>=7.4.0
pytest-cov>=4.1.0
python-multipart>=0.0.6
pydantic>=2.0.0
# Security and compliance features
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
PyJWT>=2.8.0

# Database and ORM
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.7
databases[postgresql]>=0.8.0
asyncpg>=0.28.0

# Additional production features
bcrypt>=4.0.1
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
redis>=4.6.0
